Merge _http.py back into _urllib2_fork.py
This net result of the import from Python trunk and this merge back from
mechanize is to merge a bunch of changes from Python trunk.

This commit merges only the code that's common to _http.py and urllib2
(other modules to follow shortly).

This commit merges only the "obvious" changes from Python trunk (most of
it).  Other changes from trunk will be merged after the initial re-sync.
Some bad changes from Python trunk will not be merged.  Some changes will
be merged in a later release.